requirements.txt
# Core dependencies
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# Data handling and storage
sqlite3
json5==0.9.14

# Visualization and analysis
matplotlib==3.8.2
seaborn==0.13.0

# Async support
asyncio
aiohttp==3.9.1
websockets==12.0

# Logging and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3